add Profile/UpdateCreditsField action

+++ Datas/AddedCreditsData.php
@@ -11,5 +11,5 @@ class AddedCreditsData extends Data
     // public string $adminId;
     public string $profileId;
     public string $userId;
-    public int $credit;
+    public float $credit;
 }
